% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kinesis_operations.R
\name{describe_limits}
\alias{describe_limits}
\title{Describes the shard limits and usage for the account}
\usage{
describe_limits()
}
\description{
Describes the shard limits and usage for the account.
}
\details{
If you update your account limits, the old limits might be returned for a few minutes.

This operation has a limit of one transaction per second per account.
}
\section{Accepted Parameters}{
\preformatted{describe_limits()
}
}

