% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kinesis_operations.R
\name{describe_stream_summary}
\alias{describe_stream_summary}
\title{Provides a summarized description of the specified Kinesis data stream without the shard list}
\usage{
describe_stream_summary(
  StreamName = "string"
)
}
\arguments{
\item{StreamName}{[required] The name of the stream to describe.}
}
\description{
Provides a summarized description of the specified Kinesis data stream without the shard list.
}
\details{
The information returned includes the stream name, Amazon Resource Name (ARN), status, record retention period, approximate creation time, monitoring, encryption details, and open shard count.
}
\examples{

}
