% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kinesis_operations.R
\name{disable_enhanced_monitoring}
\alias{disable_enhanced_monitoring}
\title{Disables enhanced monitoring}
\usage{
disable_enhanced_monitoring(StreamName, ShardLevelMetrics)
}
\arguments{
\item{StreamName}{The name of the Kinesis data stream for which to disable enhanced monitoring.}

\item{ShardLevelMetrics}{List of shard-level metrics to disable.

The following are the valid shard-level metrics. The value "\code{ALL}" disables every metric.
\itemize{
\item \code{IncomingBytes}
\item \code{IncomingRecords}
\item \code{OutgoingBytes}
\item \code{OutgoingRecords}
\item \code{WriteProvisionedThroughputExceeded}
\item \code{ReadProvisionedThroughputExceeded}
\item \code{IteratorAgeMilliseconds}
\item \code{ALL}
}

For more information, see \href{http://docs.aws.amazon.com/kinesis/latest/dev/monitoring-with-cloudwatch.html}{Monitoring the Amazon Kinesis Data Streams Service with Amazon CloudWatch} in the \emph{Amazon Kinesis Data Streams Developer Guide}.}
}
\description{
Disables enhanced monitoring.
}
\examples{

}
