% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kinesis_operations.R
\name{increase_stream_retention_period}
\alias{increase_stream_retention_period}
\title{Increases the Kinesis data stream's retention period, which is the length of time data records are accessible after they are added to the stream}
\usage{
increase_stream_retention_period(StreamName, RetentionPeriodHours)
}
\arguments{
\item{StreamName}{The name of the stream to modify.}

\item{RetentionPeriodHours}{The new retention period of the stream, in hours. Must be more than the current retention period.}
}
\description{
Increases the Kinesis data stream's retention period, which is the length of time data records are accessible after they are added to the stream. The maximum value of a stream's retention period is 168 hours (7 days).
}
\details{
If you choose a longer stream retention period, this operation increases the time period during which records that have not yet expired are accessible. However, it does not make previous, expired data (older than the stream's previous retention period) accessible after the operation has been called. For example, if a stream's retention period is set to 24 hours and is increased to 168 hours, any data that is older than 24 hours remains inaccessible to consumer applications.
}
\examples{

}
