% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kinesis_package.R
\docType{package}
\name{paws.kinesis-package}
\alias{paws.kinesis}
\alias{paws.kinesis-package}
\title{paws.kinesis: Amazon Kinesis}
\description{
Amazon Kinesis Data Streams is a managed service
that scales elastically for real-time processing of streaming big
data.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
