% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kinesis_operations.R
\name{remove_tags_from_stream}
\alias{remove_tags_from_stream}
\title{Removes tags from the specified Kinesis data stream}
\usage{
remove_tags_from_stream(
  StreamName = "string",
  TagKeys = list(
    "string"
  )
)
}
\arguments{
\item{StreamName}{[required] The name of the stream.}

\item{TagKeys}{[required] A list of tag keys. Each corresponding tag is removed from the stream.}
}
\description{
Removes tags from the specified Kinesis data stream. Removed tags are deleted and cannot be recovered after this operation successfully completes.
}
\details{
If you specify a tag that does not exist, it is ignored.

RemoveTagsFromStream has a limit of five transactions per second per account.
}
\examples{

}
